Add dicom_sniper
- dicom_net_series_finder is superseded by dicom_sniper
- upload and dicom functionality factored out into separate files
- reapers and folder uploader adpated to new functions
- more renaming to come
